+++ apps/block_scout_web/test/block_scout_web/controllers/address_write_proxy_controller_test.exs
@@ -43,7 +43,7 @@ defmodule BlockScoutWeb.AddressWriteProxyControllerTest do
         block_index: 0
       )
 
-      insert(:smart_contract, address_hash: contract_address.hash)
+      insert(:smart_contract, address_hash: contract_address.hash, contract_code_md5: "123")
 
       get_eip1967_implementation()