+++ apps/block_scout_web/test/block_scout_web/controllers/api/rpc/contract_controller_test.exs
@@ -35,7 +35,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
     end
 
     test "with a verified smart contract, all contract information is shown", %{conn: conn, params: params} do
-      contract = insert(:smart_contract)
+      contract = insert(:smart_contract, contract_code_md5: "123")
 
       response =
         conn
@@ -81,7 +81,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
 
     test "filtering for only unverified contracts shows only unverified contracts", %{params: params, conn: conn} do
       address = insert(:contract_address)
-      insert(:smart_contract)
+      insert(:smart_contract, contract_code_md5: "123")
 
       response =
         conn
@@ -106,7 +106,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
       conn: conn
     } do
       address = insert(:contract_address)
-      insert(:smart_contract)
+      insert(:smart_contract, contract_code_md5: "123")
       insert(:contract_address, contract_code: "0x")
 
       response =
@@ -129,7 +129,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
 
     test "filtering for only verified contracts shows only verified contracts", %{params: params, conn: conn} do
       insert(:contract_address)
-      contract = insert(:smart_contract)
+      contract = insert(:smart_contract, contract_code_md5: "123")
 
       response =
         conn
@@ -221,7 +221,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
 
     test "filtering for only not_decompiled (and by extension not verified contracts)", %{params: params, conn: conn} do
       insert(:decompiled_smart_contract)
-      insert(:smart_contract)
+      insert(:smart_contract, contract_code_md5: "123")
       contract_address = insert(:contract_address)
 
       response =
@@ -247,7 +247,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
       conn: conn
     } do
       insert(:decompiled_smart_contract)
-      insert(:smart_contract)
+      insert(:smart_contract, contract_code_md5: "123")
       insert(:contract_address, contract_code: "0x")
       contract_address = insert(:contract_address)
 
@@ -327,7 +327,7 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
     end
 
     test "with a verified contract address", %{conn: conn} do
-      contract = insert(:smart_contract)
+      contract = insert(:smart_contract, contract_code_md5: "123")
 
       params = %{
         "module" => "contract",
@@ -423,7 +423,13 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
     end
 
     test "with a verified contract address", %{conn: conn} do
-      contract = insert(:smart_contract, optimization: true, optimization_runs: 200, evm_version: "default")
+      contract =
+        insert(:smart_contract,
+          optimization: true,
+          optimization_runs: 200,
+          evm_version: "default",
+          contract_code_md5: "123"
+        )
 
       params = %{
         "module" => "contract",
@@ -469,7 +475,8 @@ defmodule BlockScoutWeb.API.RPC.ContractControllerTest do
           optimization_runs: 200,
           evm_version: "default",
           constructor_arguments:
-            "00000000000000000000000008e7592ce0d7ebabf42844b62ee6a878d4e1913e000000000000000000000000e1b6037da5f1d756499e184ca15254a981c92546"
+            "00000000000000000000000008e7592ce0d7ebabf42844b62ee6a878d4e1913e000000000000000000000000e1b6037da5f1d756499e184ca15254a981c92546",
+          contract_code_md5: "123"
         )
 
       params = %{