+++ apps/explorer/lib/explorer/chain/smart_contract.ex
@@ -198,6 +198,7 @@ defmodule Explorer.Chain.SmartContract do
   * `file_path` - show the filename or path to the file of the contract source file
   * `is_changed_bytecode` - boolean flag, determines if contract's bytecode was modified 
   * `bytecode_checked_at` - timestamp of the last check of contract's bytecode matching (DB and BlockChain)
+  * `contract_code_md5` - md5(`t:Explorer.Chain.Address.t/0` `contract_code`)
   """
 
   @type t :: %Explorer.Chain.SmartContract{
@@ -214,7 +215,8 @@ defmodule Explorer.Chain.SmartContract do
           file_path: String.t(),
           is_vyper_contract: boolean | nil,
           is_changed_bytecode: boolean,
-          bytecode_checked_at: DateTime.t()
+          bytecode_checked_at: DateTime.t(),
+          contract_code_md5: String.t()
         }
 
   schema "smart_contracts" do
@@ -234,6 +236,7 @@ defmodule Explorer.Chain.SmartContract do
     field(:is_vyper_contract, :boolean)
     field(:is_changed_bytecode, :boolean, default: false)
     field(:bytecode_checked_at, :utc_datetime_usec, default: DateTime.add(DateTime.utc_now(), -86400, :second))
+    field(:contract_code_md5, :string)
 
     has_many(
       :decompiled_smart_contracts,
@@ -273,9 +276,18 @@ defmodule Explorer.Chain.SmartContract do
       :file_path,
       :is_vyper_contract,
       :is_changed_bytecode,
-      :bytecode_checked_at
+      :bytecode_checked_at,
+      :contract_code_md5
     ])
-    |> validate_required([:name, :compiler_version, :optimization, :contract_source_code, :abi, :address_hash])
+    |> validate_required([
+      :name,
+      :compiler_version,
+      :optimization,
+      :contract_source_code,
+      :abi,
+      :address_hash,
+      :contract_code_md5
+    ])
     |> unique_constraint(:address_hash)
     |> prepare_changes(&upsert_contract_methods/1)
   end
@@ -303,11 +315,12 @@ defmodule Explorer.Chain.SmartContract do
         :file_path,
         :is_vyper_contract,
         :is_changed_bytecode,
-        :bytecode_checked_at
+        :bytecode_checked_at,
+        :contract_code_md5
       ])
       |> (&if(json_verification,
             do: &1,
-            else: validate_required(&1, [:name, :compiler_version, :optimization, :address_hash])
+            else: validate_required(&1, [:name, :compiler_version, :optimization, :address_hash, :contract_code_md5])
           )).()
 
     field_to_put_message = if json_verification, do: :file, else: :contract_source_code