+++ apps/explorer/lib/explorer/smart_contract/solidity/publisher.ex
@@ -47,7 +47,7 @@ defmodule Explorer.SmartContract.Solidity.Publisher do
     end
   end
 
-  def publish_with_standart_json_input(%{"address_hash" => address_hash} = params, json_input) do
+  def publish_with_standard_json_input(%{"address_hash" => address_hash} = params, json_input) do
     case Verifier.evaluate_authenticity_via_standard_json_input(address_hash, params, json_input) do
       {:ok, %{abi: abi, constructor_arguments: constructor_arguments}, additional_params} ->
         params_with_constructor_arguments =